# [PERF] point_of_sale: improve get_price for categories