### Description:

When having a pricelist with a lot of categories, it slows down the frontend when `get_price` is triggered. This is caused by the fact that the commit eb805e8 pre-filter product and product template. However, it doesn't pre-filter the categories. This means that if a user has, for example, 2918 categories, it will need to filter them each time `get_price` is triggered on each product updated.

### Fix:

To fix that, we can just pre-filter categories so that when `get_price` is called, we just need to filter on the items related to the product.

When opening the PoS, `_loadProductProduct` is called to retrieve the products. With these changes, this method will be a little slower to speed up `get_price`, which is called at the loading and anytime the display updates (adding to cart, selecting other categories, etc.).

### Benchmark (min-max time):

`get_price`:
| N° of categ. |  Before  | After |
|--------------|----------|-------|
|         2918 | 19-408ms | 1-4ms |
|         1918 | 15-240ms | 1-4ms |
|          518 | 14-181ms | 1-4ms |

`_loadProductProduct`:
| N° of categ. |  Before  |   After   |
|--------------|----------|-----------|
|         2918 |  6-734ms | 56ms-1.4s |
|         1918 |  9-632ms | 35ms-1.2s |
|          518 |  9-331ms |  44-827ms |

### References:
eb805e85f4a3bb7113c67feeb1d7101714bd4238
opw-4520767