@robodoo 2025-02-14
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/197799.png)](https://mergebot.odoo.com/odoo/odoo/pull/197799)