@robodoo 2025-02-14
Forward-porting to '17.0'.