@davidmonnom 2025-02-17
Hey, thanks for the PR, may I know with how many products you benchmarked?

I think it's possible to reduce the speed of `_loadProductProduct` but depending on the question above