@davidmonnom 2025-02-18
Hello @auma-odoo I'm currently waiting advice from @caburj