@auma-odoo 2025-02-18
@davidmonnom @caburj Should be good now